+++ docs/source/racs/changes.rst
@@ -25,9 +25,16 @@ Effect
 ^^^^^^
 * ``is_specializable`` looks inside range.
 
+
+Components
+^^^^^^^^^^
+* ALP implemented
+* Custom mutation in GA
+
 Thoughts
 ^^^^^^^^
 * Maybe effect could return just encoded value, instead of UBR...
 * Specificity/generality should measure how wide is the UBR
 * Maybe ``u_max`` should hold information how specific condition should be (not just wildcards but spread)
 * In the end of ALP phase we should perform something like classifier merge
+* There is still very aggressive``generalize`` function in Condition part.